hydrate write.rs into reply.rs
